\name{QueftsModel-class}

\docType{class}

\alias{QueftsCrop}
\alias{QueftsModel}
\alias{QueftsSoil}

\alias{QueftsCrop-class}
\alias{QueftsModel-class}
\alias{QueftsSoil-class}

\alias{Rcpp_QueftsCrop-class}
\alias{Rcpp_QueftsModel-class}
\alias{Rcpp_QueftsSoil-class}

\alias{show,Rcpp_QueftsModel-method}


\title{QueftsModel class}

\description{
These are the classes!
}


\section{Objects from the Class}{
QueftsModel$new()
}



\keyword{classes}
\keyword{spatial}
