\name{Rquefts-package}

\alias{Rquefts-package}

\alias{Rquefts}

\docType{package}

\title{
\packageTitle{Rquefts}
}

\description{

This package provides a R interface to a C++ implementation of the QUEFTS model.
QUEFTS (Quantitative Evaluation of the Native Fertility of Tropical Soils) model (1) estimates native nutrient (N, P, K) supply of soils from a few soil chemical properties; and (2) computes crop yield given that supply, fertilizer application and crop parameters.  See Janssen et al. (1990) <doi:10.1016/0016-7061(90)90021-Z> for the technical details and Sattari et al. (2014) <doi:10.1016/j.fcr.2013.12.005> for a recent evaluation and improvements. 
}
