\name{Rquefts-package}

\alias{Rquefts-package}

\alias{Rquefts}

\docType{package}

\title{
\packageTitle{Rquefts}
}

\description{

This package provides implments the QUEFTS model.

QUEFTS (Quantitative Evaluation of the Native Fertility of Tropical Soils) model (1) estimates native nutrient (N, P, K) supply of soils from a few soil chemical properties; and (2) computes crop yield given that supply, fertilizer application and crop parameters.  See Janssen et al. (1990) <doi:10.1016/0016-7061(90)90021-Z> for the technical details and Sattari et al. (2014) <doi:10.1016/j.fcr.2013.12.005> for a recent evaluation and improvements. 

The package is particularly useful if you want to make spatial predictions with QUEFTS.

There are also a few functions that can help with computing the amount of nutriets supplied with fertilizer (blends) and compute the optimal use of fertilizer given a goal in nutrients, available products, and their prices.

}
